+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/association.ts
@@ -32,6 +32,7 @@ export interface MyWidgetPreviewProps {
     referenceOrSet: string;
     optionsSource: {} | { type: string } | null;
     displayValue: string;
+    referenceOnChange: {} | null;
 }
 `;